+++ python/CHANGELOG.rst
@@ -17,12 +17,27 @@
   so a set of tree sequence can be added to the right of an existing one.
   (:user:`hyanwong`, :pr:`3165`, :issue:`3164`)
 
+- Add ``TreeSequence.map_to_vcf_model`` method to return a mapping of
+  the tree sequence to the VCF model. 
+  (:user:`benjeffery`, :pr:`3163`)
 
 **Fixes**
 
 - Correct assertion message when tables are compared with metadata ignored.
   (:user:`benjeffery`, :pr:`3162`, :issue:`3161`)
   
+**Breaking changes** 
+
+- ``TreeSequence.write_vcf`` now filters non-sample nodes from individuals
+  by default, instead of raising an error. These nodes can be included using the
+  new ``include_non_sample_nodes`` argument. 
+  By default individual names (sample IDs) in VCF output are now of the form
+  ``tsk_{individual.id}`` Previously these were always 
+  ``"tsk_{j}" for j in range(num_individuals)``. This may break some downstream
+  code if individuals are specified. To fix, manually specify ``individual_names``
+  to the required pattern.
+  (:user:`benjeffery`, :pr:`3163`)
+
 --------------------
 [0.6.3] - 2025-04-28
 --------------------